+++ server/async_tcp.go
@@ -1,15 +1,11 @@
 package server
 
 import (
-	"log"
-	"net"
 	"os"
 	"sync"
 	"sync/atomic"
-	"syscall"
 	"time"
 
-	"github.com/dicedb/dice/config"
 	"github.com/dicedb/dice/core"
 )
 
@@ -48,144 +44,4 @@ func WaitForSignal(wg *sync.WaitGroup, sigs chan os.Signal) {
 	// if server is in any other state, initiate a shutdown
 	core.Shutdown()
 	os.Exit(0)
-}
-
-func RunAsyncTCPServer(wg *sync.WaitGroup) error {
-	defer wg.Done()
-	defer func() {
-		atomic.StoreInt32(&eStatus, EngineStatus_SHUTTING_DOWN)
-	}()
-
-	log.Println("starting an asynchronous TCP server on", config.Host, config.Port)
-
-	max_clients := 20000
-
-	// Create EPOLL Event Objects to hold events
-	var events []syscall.EpollEvent = make([]syscall.EpollEvent, max_clients)
-
-	// Create a socket
-	serverFD, err := syscall.Socket(syscall.AF_INET, syscall.O_NONBLOCK|syscall.SOCK_STREAM, 0)
-	if err != nil {
-		return err
-	}
-	defer syscall.Close(serverFD)
-
-	// Set the Socket operate in a non-blocking mode
-	if err = syscall.SetNonblock(serverFD, true); err != nil {
-		return err
-	}
-
-	// Bind the IP and the port
-	ip4 := net.ParseIP(config.Host)
-
-	if err = syscall.Bind(serverFD, &syscall.SockaddrInet4{
-		Port: config.Port,
-		Addr: [4]byte{ip4[0], ip4[1], ip4[2], ip4[3]},
-	}); err != nil {
-		return err
-	}
-
-	// Start listening
-	if err = syscall.Listen(serverFD, max_clients); err != nil {
-		return err
-	}
-
-	// AsyncIO starts here!!
-
-	// creating EPOLL instance
-	epollFD, err := syscall.EpollCreate1(0)
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer syscall.Close(epollFD)
-
-	// Specify the events we want to get hints about
-	// and set the socket on which
-	var socketServerEvent syscall.EpollEvent = syscall.EpollEvent{
-		Events: syscall.EPOLLIN,
-		Fd:     int32(serverFD),
-	}
-
-	// Listen to read events on the Server itself
-	if err = syscall.EpollCtl(epollFD, syscall.EPOLL_CTL_ADD, serverFD, &socketServerEvent); err != nil {
-		return err
-	}
-
-	// loop until the server is not shutting down
-	for atomic.LoadInt32(&eStatus) != EngineStatus_SHUTTING_DOWN {
-		if time.Now().After(lastCronExecTime.Add(cronFrequency)) {
-			core.DeleteExpiredKeys()
-			lastCronExecTime = time.Now()
-		}
-
-		// Say, the Engine triggered SHUTTING down when the control flow is here ->
-		// Current: Engine status == WAITING
-		// Update: Engine status = SHUTTING_DOWN
-		// Then we have to exit (handled in Signal Handler)
-
-		// see if any FD is ready for an IO
-		nevents, e := syscall.EpollWait(epollFD, events[:], -1)
-		if e != nil {
-			continue
-		}
-
-		// Here, we do not want server to go back from SHUTTING DOWN
-		// to BUSY
-		// If the engine status == SHUTTING_DOWN over here ->
-		// We have to exit
-		// hence the only legal transitiion is from WAITING to BUSY
-		// if that does not happen then we can exit.
-
-		// mark engine as BUSY only when it is in the waiting state
-		if !atomic.CompareAndSwapInt32(&eStatus, EngineStatus_WAITING, EngineStatus_BUSY) {
-			// if swap unsuccessful then the existing status is not WAITING, but something else
-			switch eStatus {
-			case EngineStatus_SHUTTING_DOWN:
-				return nil
-			}
-		}
-
-		for i := 0; i < nevents; i++ {
-			// if the socket server itself is ready for an IO
-			if int(events[i].Fd) == serverFD {
-				// accept the incoming connection from a client
-				fd, _, err := syscall.Accept(serverFD)
-				if err != nil {
-					log.Println("err", err)
-					continue
-				}
-
-				connectedClients[fd] = core.NewClient(fd)
-				syscall.SetNonblock(fd, true)
-
-				// add this new TCP connection to be monitored
-				var socketClientEvent syscall.EpollEvent = syscall.EpollEvent{
-					Events: syscall.EPOLLIN,
-					Fd:     int32(fd),
-				}
-				if err := syscall.EpollCtl(epollFD, syscall.EPOLL_CTL_ADD, fd, &socketClientEvent); err != nil {
-					log.Fatal(err)
-				}
-			} else {
-				comm := connectedClients[int(events[i].Fd)]
-				if comm == nil {
-					continue
-				}
-				cmds, maliciousFlag, err := readCommands(comm)
-				if err != nil || maliciousFlag {
-					syscall.Close(int(events[i].Fd))
-					delete(connectedClients, int(events[i].Fd))
-					continue
-				}
-				respond(cmds, comm)
-			}
-		}
-
-		// mark engine as WAITING
-		// no contention as the signal handler is blocked until
-		// the engine is BUSY
-		atomic.StoreInt32(&eStatus, EngineStatus_WAITING)
-	}
-
-	return nil
 }
